pyproject.toml
[tool.poetry]
name = "learnfastapi"
version = "0.1.0"
description = ""
authors = ["M"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
click = "^8.1.3"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
uvicorn = "^0.21.1"
SQLAlchemy = "^2.0.9"
ipython = "^8.12.0"
fastapi-debug-toolbar = "^0.3.2"
pytest = "^7.3.0"
httpx = "^0.24.0"
py-spy = "^0.3.14"
alembic = "^1.10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
